Send player's vote confirmation to group
Whenever voting starts, players don't know if a fellow player has voted. For each voting, players have to wait for every other player to vote.
This change tries to make it easier for players to know which player hasn't voted.
The bot will send a message to the group for each player that actually votes, and players, knowing which players havn't vote, can take the proper actions to continue the game.